Projects

+++ src/app/my-projects/projects/project-item/project-item.component.ts
@@ -6,21 +6,37 @@ import {
 import { Router } from '@angular/router';
 import { Project } from '../../../shared/models/project.model';
 import { ProjectService } from '../../../shared/services/project.service';
-import {ViewEncapsulation} from '@angular/core';
+import { ViewEncapsulation } from '@angular/core';
+import { SimpleNotificationsModule, NotificationsService } from 'angular2-notifications';
+import { Modal, BSModalContext } from 'angular2-modal/plugins/bootstrap';
+import { ProjectShareModalComponent } from '../../project-share-modal/project-share-modal.component';
+import { overlayConfigFactory } from 'angular2-modal';
+import { WorkSpaceService } from '../../../shared/services/work-space.service';
+import { AuthService } from '../../../shared/services/auth.service';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 @Component({
   selector: '[app-project-item]',
   templateUrl: './project-item.component.html',
   styleUrls: ['./project-item.component.css'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
+  providers: [Modal]
 })
 export class ProjectItemComponent implements OnInit {
   @Input() project: Project;
-  editableText = 'myText';
+  public myNick = '';
+  public myNick$ = new BehaviorSubject<string>('');
 
-  constructor(private projectService: ProjectService, private router: Router) { }
+  constructor(private projectService: ProjectService, private router: Router, private notificationsService: NotificationsService,
+    private modal: Modal, private workspaceService: WorkSpaceService, private authService: AuthService) { }
 
   ngOnInit() {
+    this.authService.getEmail().subscribe(
+      nick => {
+        this.myNick = nick;
+      },
+      error => console.error(error)
+    );
   }
 
   onUpdateDescription(desc) {
@@ -28,25 +44,37 @@ export class ProjectItemComponent implements OnInit {
     this.projectService.updateProject(this.project)
       .subscribe(
       resp => {
+        this.notificationsService.success('Exito', 'Se actualizo la descripcion del proyecto ' + this.project.name);
         this.project._etag = resp._etag;
-        console.log('UPDATE:' + resp);
       },
       error => {
-        console.error(error);
+        this.notificationsService.error('Error', 'Error en la actualizacion del proyecto');
       });
   }
 
   onDeleteProject() {
-    const id = this.project._id;
-    const projToDelete = this.projectService.getProject(id);
-    this.projectService.deleteProject(projToDelete)
-      .subscribe(
-      resp => {
-        console.log('DELETE:' + resp);
-        this.projectService.removeProject(projToDelete);
-      },
-      error => {
-        console.error(error);
+    const dialogRef = this.modal.confirm().size('lg').isBlocking(true).showClose(true).keyboard(27)
+      .okBtn('Confirmar').okBtnClass('btn btn-info').cancelBtnClass('btn btn-danger')
+      .title('Eliminar proyecto').body(' Seguro que desea eliminar el proyecto y todos los documentos asociados? ').open();
+    dialogRef
+      .then(r => {
+        r.result
+          .then(result => {
+            const id = this.project._id;
+            const projToDelete = this.projectService.getProject(id);
+            this.projectService.deleteProject(projToDelete)
+              .subscribe(
+              resp => {
+                this.notificationsService.success('Exito', 'El proyecto se elimino correctamente');
+                this.projectService.removeProject(projToDelete);
+              },
+              error => {
+                this.notificationsService.error('Error', 'Error en el borrado del proyecto');
+              });
+          })
+          .catch(error =>
+            console.log(error)
+          );
       });
   }
 
@@ -54,4 +82,15 @@ export class ProjectItemComponent implements OnInit {
     this.router.navigate(['workspace', this.project._id]);
   }
 
+  onShareProject() {
+    const project = this.workspaceService.getProjectId();
+    this.modal.open(ProjectShareModalComponent, overlayConfigFactory({ project: this.project }, BSModalContext))
+      .then((resultPromise) => {
+        resultPromise.result.then((result) => {
+          if (result != null) {
+            this.modal.alert().headerClass('btn-danger').title('Error al guardar').body(result).open();
+          }
+        });
+      });
+  }
 }
